requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.9
flask_cors==3.0.10
requests==2.30.0
flask_socketio
pika
pymongo
flask-cors
gevent
gevent-websocket